chore(remix): split app gen to smaller steps
