+++ packages/remix/src/generators/application/lib/index.ts
@@ -1,3 +1,5 @@
 export * from './normalize-options';
 export * from './update-unit-test-config';
 export * from './add-e2e';
+export * from './add-lint';
+export * from './add-unittest';
